sentry - organize modules: - application - move Application, Session & Auth - routers - move from lib - response - ResponseError & helper fns
