+++ sentry/src/middleware/campaign.rs
@@ -1,5 +1,5 @@
 use crate::{db::fetch_campaign, middleware::Middleware};
-use crate::{Application, Auth, ResponseError, RouteParams};
+use crate::{response::ResponseError, routes::routers::RouteParams, Application, Auth};
 use adapter::client::Locked;
 use async_trait::async_trait;
 use hyper::{Body, Request};
