+++ sentry/src/middleware/channel.rs
@@ -1,5 +1,6 @@
 use crate::{
-    db::get_channel_by_id, middleware::Middleware, Application, Auth, ResponseError, RouteParams,
+    db::get_channel_by_id, middleware::Middleware, response::ResponseError,
+    routes::routers::RouteParams, Application, Auth,
 };
 use adapter::client::Locked;
 use futures::future::{BoxFuture, FutureExt};
