+++ sentry/src/routes/routers.rs
@@ -1,5 +1,8 @@
 //! This module contains all the Sentry REST API routers.
 //!
+//! Ideally these routers will be replaced with proper routing,
+//! when we replace the custom `hyper` server setup with a web framework.
+//!
 //! # Routers
 //!
 //! Routers are functions that are called on certain route prefix (e.g. `/v5/channel`, `/v5/campaign`)
@@ -12,16 +15,290 @@
 use crate::{
     middleware::{
         auth::{AuthRequired, IsAdmin},
+        campaign::{CalledByCreator, CampaignLoad},
+        channel::ChannelLoad,
         Chain, Middleware,
     },
-    routes::get_analytics,
-    Application, Auth, ResponseError,
+    response::ResponseError,
+    routes::{
+        analytics::analytics,
+        campaign,
+        campaign::{campaign_list, create_campaign, update_campaign},
+        channel::{
+            add_spender_leaf, channel_list, channel_payout, get_accounting_for_channel,
+            get_all_spender_limits, get_spender_limits, last_approved,
+            validator_message::{
+                create_validator_messages, extract_params, list_validator_messages,
+            },
+        },
+    },
+    Application, Auth,
 };
 use adapter::prelude::*;
 use hyper::{Body, Method, Request, Response};
+use once_cell::sync::Lazy;
 use primitives::analytics::{query::AllowedKey, AuthenticateAs};
+use regex::Regex;
 
 use super::units_for_slot::post_units_for_slot;
+
+pub(crate) static LAST_APPROVED_BY_CHANNEL_ID: Lazy<Regex> = Lazy::new(|| {
+    Regex::new(r"^/v5/channel/0x([a-zA-Z0-9]{64})/last-approved/?$")
+        .expect("The regex should be valid")
+});
+
+/// Only the initial Regex to be matched.
+pub(crate) static CHANNEL_VALIDATOR_MESSAGES: Lazy<Regex> = Lazy::new(|| {
+    Regex::new(r"^/v5/channel/0x([a-zA-Z0-9]{64})/validator-messages(/.*)?$")
+        .expect("The regex should be valid")
+});
+pub(crate) static CHANNEL_SPENDER_LEAF_AND_TOTAL_DEPOSITED: Lazy<Regex> = Lazy::new(|| {
+    Regex::new(r"^/v5/channel/0x([a-zA-Z0-9]{64})/spender/0x([a-zA-Z0-9]{40})/?$")
+        .expect("This regex should be valid")
+});
+
+pub(crate) static INSERT_EVENTS_BY_CAMPAIGN_ID: Lazy<Regex> = Lazy::new(|| {
+    Regex::new(r"^/v5/campaign/(0x[a-zA-Z0-9]{32})/events/?$").expect("The regex should be valid")
+});
+
+pub(crate) static CLOSE_CAMPAIGN_BY_CAMPAIGN_ID: Lazy<Regex> = Lazy::new(|| {
+    Regex::new(r"^/v5/campaign/(0x[a-zA-Z0-9]{32})/close/?$").expect("The regex should be valid")
+});
+
+pub(crate) static CAMPAIGN_UPDATE_BY_ID: Lazy<Regex> = Lazy::new(|| {
+    Regex::new(r"^/v5/campaign/(0x[a-zA-Z0-9]{32})/?$").expect("The regex should be valid")
+});
+pub(crate) static CHANNEL_ALL_SPENDER_LIMITS: Lazy<Regex> = Lazy::new(|| {
+    Regex::new(r"^/v5/channel/0x([a-zA-Z0-9]{64})/spender/all/?$")
+        .expect("The regex should be valid")
+});
+pub(crate) static CHANNEL_ACCOUNTING: Lazy<Regex> = Lazy::new(|| {
+    Regex::new(r"^/v5/channel/0x([a-zA-Z0-9]{64})/accounting/?$")
+        .expect("The regex should be valid")
+});
+pub(crate) static CHANNEL_PAY: Lazy<Regex> = Lazy::new(|| {
+    Regex::new(r"^/v5/channel/0x([a-zA-Z0-9]{64})/pay/?$").expect("The regex should be valid")
+});
+
+/// Regex extracted parameters.
+/// This struct is created manually on each of the matched routes.
+#[derive(Debug, Clone)]
+pub struct RouteParams(pub Vec<String>);
+
+impl RouteParams {
+    pub fn get(&self, index: usize) -> Option<String> {
+        self.0.get(index).map(ToOwned::to_owned)
+    }
+
+    pub fn index(&self, i: usize) -> String {
+        self.0[i].clone()
+    }
+}
+
+// TODO AIP#61: Add routes for:
+// - GET /channel/:id/get-leaf
+pub async fn channels_router<C: Locked + 'static>(
+    mut req: Request<Body>,
+    app: &Application<C>,
+) -> Result<Response<Body>, ResponseError> {
+    let (path, method) = (req.uri().path().to_owned(), req.method());
+
+    // `GET /v5/channel/list`
+    if let ("/v5/channel/list", &Method::GET) = (path.as_str(), method) {
+        channel_list(req, app).await
+    }
+    // `GET /v5/channel/:id/last-approved`
+    else if let (Some(caps), &Method::GET) = (LAST_APPROVED_BY_CHANNEL_ID.captures(&path), method)
+    {
+        let param = RouteParams(vec![caps
+            .get(1)
+            .map_or("".to_string(), |m| m.as_str().to_string())]);
+        req.extensions_mut().insert(param);
+
+        req = ChannelLoad.call(req, app).await?;
+
+        last_approved(req, app).await
+    }
+    // `GET /v5/channel/:id/validator-messages`
+    else if let (Some(caps), &Method::GET) = (CHANNEL_VALIDATOR_MESSAGES.captures(&path), method)
+    {
+        let param = RouteParams(vec![caps
+            .get(1)
+            .map_or("".to_string(), |m| m.as_str().to_string())]);
+
+        req.extensions_mut().insert(param);
+
+        req = ChannelLoad.call(req, app).await?;
+
+        // @TODO: Move this to a middleware?!
+        let extract_params = match extract_params(caps.get(2).map_or("", |m| m.as_str())) {
+            Ok(params) => params,
+            Err(error) => {
+                return Err(error.into());
+            }
+        };
+
+        list_validator_messages(req, app, &extract_params.0, &extract_params.1).await
+    }
+    // `POST /v5/channel/:id/validator-messages`
+    else if let (Some(caps), &Method::POST) = (CHANNEL_VALIDATOR_MESSAGES.captures(&path), method)
+    {
+        let param = RouteParams(vec![caps
+            .get(1)
+            .map_or("".to_string(), |m| m.as_str().to_string())]);
+
+        req.extensions_mut().insert(param);
+
+        let req = Chain::new()
+            .chain(AuthRequired)
+            .chain(ChannelLoad)
+            .apply(req, app)
+            .await?;
+
+        create_validator_messages(req, app).await
+    }
+    // `GET /v5/channel/:id/spender/:addr`
+    else if let (Some(caps), &Method::GET) = (
+        CHANNEL_SPENDER_LEAF_AND_TOTAL_DEPOSITED.captures(&path),
+        method,
+    ) {
+        let param = RouteParams(vec![
+            caps.get(1)
+                .map_or("".to_string(), |m| m.as_str().to_string()), // channel ID
+            caps.get(2)
+                .map_or("".to_string(), |m| m.as_str().to_string()), // spender addr
+        ]);
+        req.extensions_mut().insert(param);
+        req = Chain::new()
+            .chain(AuthRequired)
+            .chain(ChannelLoad)
+            .apply(req, app)
+            .await?;
+
+        get_spender_limits(req, app).await
+    }
+    // `POST /v5/channel/:id/spender/:addr`
+    else if let (Some(caps), &Method::POST) = (
+        CHANNEL_SPENDER_LEAF_AND_TOTAL_DEPOSITED.captures(&path),
+        method,
+    ) {
+        let param = RouteParams(vec![
+            caps.get(1)
+                .map_or("".to_string(), |m| m.as_str().to_string()), // channel ID
+            caps.get(2)
+                .map_or("".to_string(), |m| m.as_str().to_string()), // spender addr
+        ]);
+        req.extensions_mut().insert(param);
+        req = Chain::new()
+            .chain(AuthRequired)
+            .chain(ChannelLoad)
+            .apply(req, app)
+            .await?;
+
+        add_spender_leaf(req, app).await
+    }
+    // `GET /v5/channel/:id/spender/all`
+    else if let (Some(caps), &Method::GET) = (CHANNEL_ALL_SPENDER_LIMITS.captures(&path), method)
+    {
+        let param = RouteParams(vec![caps
+            .get(1)
+            .map_or("".to_string(), |m| m.as_str().to_string())]);
+        req.extensions_mut().insert(param);
+
+        req = Chain::new()
+            .chain(AuthRequired)
+            .chain(ChannelLoad)
+            .apply(req, app)
+            .await?;
+
+        get_all_spender_limits(req, app).await
+    }
+    // `GET /v5/channel/:id/accounting`
+    else if let (Some(caps), &Method::GET) = (CHANNEL_ACCOUNTING.captures(&path), method) {
+        let param = RouteParams(vec![caps
+            .get(1)
+            .map_or("".to_string(), |m| m.as_str().to_string())]);
+        req.extensions_mut().insert(param);
+
+        req = Chain::new()
+            .chain(AuthRequired)
+            .chain(ChannelLoad)
+            .apply(req, app)
+            .await?;
+
+        get_accounting_for_channel(req, app).await
+    }
+    // POST /v5/channel/:id/pay
+    else if let (Some(caps), &Method::POST) = (CHANNEL_PAY.captures(&path), method) {
+        let param = RouteParams(vec![caps
+            .get(1)
+            .map_or("".to_string(), |m| m.as_str().to_string())]);
+        req.extensions_mut().insert(param);
+
+        req = Chain::new()
+            .chain(AuthRequired)
+            .chain(ChannelLoad)
+            .apply(req, app)
+            .await?;
+
+        channel_payout(req, app).await
+    } else {
+        Err(ResponseError::NotFound)
+    }
+}
+
+pub async fn campaigns_router<C: Locked + 'static>(
+    mut req: Request<Body>,
+    app: &Application<C>,
+) -> Result<Response<Body>, ResponseError> {
+    let (path, method) = (req.uri().path(), req.method());
+
+    // For creating campaigns
+    if (path, method) == ("/v5/campaign", &Method::POST) {
+        let req = AuthRequired.call(req, app).await?;
+
+        create_campaign(req, app).await
+    } else if let (Some(_caps), &Method::POST) = (CAMPAIGN_UPDATE_BY_ID.captures(path), method) {
+        let req = Chain::new()
+            .chain(AuthRequired)
+            .chain(CampaignLoad)
+            .chain(CalledByCreator)
+            .apply(req, app)
+            .await?;
+
+        update_campaign::handle_route(req, app).await
+    } else if let (Some(caps), &Method::POST) =
+        (INSERT_EVENTS_BY_CAMPAIGN_ID.captures(path), method)
+    {
+        let param = RouteParams(vec![caps
+            .get(1)
+            .map_or("".to_string(), |m| m.as_str().to_string())]);
+        req.extensions_mut().insert(param);
+
+        let req = CampaignLoad.call(req, app).await?;
+
+        campaign::insert_events::handle_route(req, app).await
+    } else if let (Some(caps), &Method::POST) =
+        (CLOSE_CAMPAIGN_BY_CAMPAIGN_ID.captures(path), method)
+    {
+        let param = RouteParams(vec![caps
+            .get(1)
+            .map_or("".to_string(), |m| m.as_str().to_string())]);
+        req.extensions_mut().insert(param);
+
+        req = Chain::new()
+            .chain(AuthRequired)
+            .chain(CampaignLoad)
+            .apply(req, app)
+            .await?;
+
+        campaign::close_campaign(req, app).await
+    } else if method == Method::GET && path == "/v5/campaign/list" {
+        campaign_list(req, app).await
+    } else {
+        Err(ResponseError::NotFound)
+    }
+}
 
 pub async fn units_for_slot_router<C: Locked + 'static>(
     req: Request<Body>,
@@ -48,7 +325,7 @@ pub async fn analytics_router<C: Locked + 'static>(
             let allowed_keys_for_request = vec![AllowedKey::Country, AllowedKey::AdSlotType]
                 .into_iter()
                 .collect();
-            get_analytics(req, app, Some(allowed_keys_for_request), None).await
+            analytics(req, app, Some(allowed_keys_for_request), None).await
         }
         ("/v5/analytics/for-advertiser", &Method::GET) => {
             let req = AuthRequired.call(req, app).await?;
@@ -59,7 +336,7 @@ pub async fn analytics_router<C: Locked + 'static>(
                 .map(|auth| AuthenticateAs::Advertiser(auth.uid))
                 .ok_or(ResponseError::Unauthorized)?;
 
-            get_analytics(req, app, None, Some(authenticate_as)).await
+            analytics(req, app, None, Some(authenticate_as)).await
         }
         ("/v5/analytics/for-publisher", &Method::GET) => {
             let authenticate_as = req
@@ -69,7 +346,7 @@ pub async fn analytics_router<C: Locked + 'static>(
                 .ok_or(ResponseError::Unauthorized)?;
 
             let req = AuthRequired.call(req, app).await?;
-            get_analytics(req, app, None, Some(authenticate_as)).await
+            analytics(req, app, None, Some(authenticate_as)).await
         }
         ("/v5/analytics/for-admin", &Method::GET) => {
             req = Chain::new()
@@ -77,7 +354,7 @@ pub async fn analytics_router<C: Locked + 'static>(
                 .chain(IsAdmin)
                 .apply(req, app)
                 .await?;
-            get_analytics(req, app, None, None).await
+            analytics(req, app, None, None).await
         }
         _ => Err(ResponseError::NotFound),
     }
@@ -87,8 +364,9 @@ pub async fn analytics_router<C: Locked + 'static>(
 mod analytics_router_test {
     use crate::{
         db::{analytics::update_analytics, DbPool},
+        response::ResponseError,
         test_util::setup_dummy_app,
-        Auth, ResponseError,
+        Auth,
     };
 
     use super::*;
